added same things as jaMME has from 25th October to 6th February to joMME

+++ CODE-mp/ui/ui_shared.h
@@ -397,8 +397,8 @@ typedef struct {
   float			bias;
   int				realTime;
   int				frameTime;
-	int				cursorx;
-	int				cursory;
+	float			cursorx;
+	float			cursory;
 	qboolean	debug;
 
   cachedAssets_t Assets;
